+++ tests/test_main.py
@@ -51,7 +51,7 @@ def test_NicfitApp_cookiecutter_exception(tmpdir):
         assert sysexit.value.code == 1
 
 
-@pytest.mark.skip(reason="Failing on Travis-CI only, most likely to env diffs")
+#@pytest.mark.skip(reason="Failing on Travis-CI only, most likely to env diffs")
 def test_NicfitApp_cookiecutter_real(tmpdir, unshallowed_repo):
     with pytest.raises(SystemExit) as sysexit:
         app.run(["cookiecutter", str(tmpdir), "--no-input"])